+++ lib/main.dart
@@ -1,20 +1,9 @@
-import 'package:money_manager/widgets/dialogs.dart';
-import 'package:provider/provider.dart';
 import 'package:flutter/material.dart';
 
-import 'package:money_manager/routes/balance.dart';
-import 'package:money_manager/routes/borrow.dart';
-import 'package:money_manager/routes/debt.dart';
+import 'package:money_manager/widgets/data/future_save_data_read.dart';
+import 'package:money_manager/widgets/data/main_state_providers.dart';
 
-import 'package:money_manager/types/savedata.dart';
-import 'package:money_manager/types/finance.dart';
-
-import 'package:money_manager/widgets/future_save_data_read.dart';
-import 'package:money_manager/widgets/main_navigation_bar.dart';
-import 'package:money_manager/widgets/main_state_providers.dart';
-
-import 'package:money_manager/finance.dart';
-import 'package:money_manager/savedata.dart';
+import 'package:money_manager/widgets/navigation/main_home_page.dart';
 
 void main() {
   runApp(const MyApp());
@@ -35,73 +24,5 @@ class MyApp extends StatelessWidget {
                       primarySwatch: Colors.green, useMaterial3: true),
                   home: const MyHomePage(),
                 )));
-  }
-}
-
-class MyHomePage extends StatefulWidget {
-  const MyHomePage({super.key});
-
-  @override
-  State<MyHomePage> createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  int selectedIndex = 0;
-
-  @override
-  Widget build(BuildContext context) {
-    FinancialModel financialModel =
-        Provider.of<FinancialModel>(context, listen: false);
-    DebtModel debtModel = Provider.of<DebtModel>(context, listen: false);
-    BorrowModel borrowModel = Provider.of<BorrowModel>(context, listen: false);
-    // TODO: not too sure about this
-    saveData.write(ConfigModel(
-        financialMoves: financialModel.items,
-        debtMoves: debtModel.items,
-        borrowMoves: borrowModel.items));
-
-    List<Widget> widgetOptions = [
-      const BalanceRoute(),
-      const DebtRoute(),
-      const BorrowRoute()
-    ];
-    List<String> dialogTitleOptions = [
-      "Add/subtract balance",
-      "Debt",
-      "Borrow"
-    ];
-    List<Model> stateOptions = [
-      Provider.of<FinancialModel>(context, listen: false),
-      Provider.of<DebtModel>(context, listen: false),
-      Provider.of<BorrowModel>(context, listen: false),
-    ];
-
-    return Scaffold(
-        appBar: AppBar(
-          title: const Text("Financial Moves"),
-        ),
-        // Floating action button that pops a dialog with many actions (add balance, substract balance, etc.)
-        floatingActionButton: FloatingActionButton(
-          onPressed: () {
-            var provider = stateOptions.elementAt(selectedIndex);
-            showMoveDialog(context,
-                title: dialogTitleOptions.elementAt(selectedIndex),
-                onDone: ({required balance, required descriptor}) {
-              Move item = Move(balance: balance, descriptor: descriptor);
-              provider.add(item);
-            });
-          },
-          tooltip: 'Make a change',
-          child: const Icon(Icons.add),
-        ),
-        body: widgetOptions.elementAt(selectedIndex),
-        bottomNavigationBar: MainNavigationBar(
-          selectedIndex: selectedIndex,
-          onDestinationSelected: (index) {
-            setState(() {
-              selectedIndex = index;
-            });
-          },
-        ));
   }
 }
