+++ lib/widgets/routes/borrow.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 import 'package:money_manager/types/finance.dart';
-import 'package:money_manager/finance.dart';
+import 'package:money_manager/data/finance.dart';
 
 class BorrowRoute extends StatefulWidget {
   const BorrowRoute({super.key});
